% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bandwidthC}
\alias{bandwidthC}
\title{A Function to compute the bandwidth using Rcpp}
\usage{
bandwidthC(X, n)
}
\arguments{
\item{X}{a matrix contains multivariate time series data. The rows are time points and the columns are variables.}

\item{n}{the number of time points}
}
\value{
the value of bandwidth for the data in X.
}
\description{
Compute the bandwidth using Rcpp. For a time series dataset, the bandwidth is the median Euclidean distance for the data corresponding to all time points.
}
\examples{
\dontrun{
data(mydata)
attach(mydata)
N <- length(mydata[,1])
h <- bandwidthC(cordata,N)
h
}
}
