% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CorrelationC}
\alias{CorrelationC}
\title{Running correlations function using Rcpp}
\usage{
CorrelationC(X, N, v, w)
}
\arguments{
\item{X}{a matrix contains multivariate time series data. The rows are time points and the columns are variables.}

\item{N}{the number of time points in X}

\item{v}{the number of variables in X}

\item{w}{the width of windows. Compute the running correlations in a window.}
}
\value{
a matrix of the running correlations for all variables at these time points.
}
\description{
Compute the running correlations using Rcpp
}
\examples{
\dontrun{
data(mydata)
attach(mydata)
N <- length(mydata[,1])
v <- length(mydata[1,])
w <- 15
cordata <- CorrelationR(mydata,N,v,w)
cordata
}
}
