% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCPpermR.R
\name{Max_Rmindrop}
\alias{Max_Rmindrop}
\title{The maximum drop in the minimum of the average within-phase variance}
\usage{
Max_Rmindrop(X, n, K)
}
\arguments{
\item{X}{a matrix contains multivariate time series data. The rows are time points and the columns are variables.}

\item{n}{the number of time points in X}

\item{K}{the number of change points}
}
\value{
the value of maximum drop variance
}
\description{
Given the number of change points, K, compute the maximum drop in the minimum of the average within-phase variance for each change point added.
}
