% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{medianC}
\alias{medianC}
\title{Median function using Rcpp}
\usage{
medianC(A)
}
\arguments{
\item{A}{a NumericVector}
}
\value{
the median of the values in the NumericVector A.
}
\description{
Compute the median of the values in a NumericVector using Rcpp
}
\examples{
\dontrun{
A <- c(2,1,4,5)
medianC(A)
}
}
