% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCPpermR.R
\name{Min_Rexpr}
\alias{Min_Rexpr}
\title{Minimize the average within-phase variance}
\usage{
Min_Rexpr(X, n, K)
}
\arguments{
\item{X}{a matrix contains multivariate time series data. The rows are time points and the columns are variables.}

\item{n}{the number of time points in X}

\item{K}{the number of change points}
}
\value{
the minimum of the average within-phase variance
}
\description{
Compute the minimum of the average within-phase variance with the known number of change points and the unknown locations of them.
}
