% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCPpermR.R
\name{mydata}
\alias{mydata}
\title{A illustration dataset}
\description{
A dataset used to illustrate the performance of the KCP permutation test.
}
\examples{
\dontrun{
data(mydata)
attach(mydata)
N <- length(mydata[,1])
v <- length(mydata[1,])
w <- 15
p_VarKCP <- VarKCP(mydata,N,v,w)
p_VarKCP
p_Vardrop <- Vardrop(mydata,N,v,w,2)
p_Vardrop
}
}
