% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCPpermR.R
\name{Rexpression}
\alias{Rexpression}
\title{The average within-phase variance}
\usage{
Rexpression(Cpoint, X, n, K)
}
\arguments{
\item{Cpoint}{contains the locations of the K change points}

\item{X}{a matrix contains multivariate time series data. The rows are time points and the columns are variables.}

\item{n}{the number of time points in X}

\item{K}{the number of change points}
}
\value{
The value of average within-phase variance for K+1 phases
}
\description{
Given the number of change points, K, compute the average within-phase variance
}
