% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SortC}
\alias{SortC}
\title{Sort function using Rcpp}
\usage{
SortC(x)
}
\arguments{
\item{x}{a NumericVector}
}
\value{
a vector contains the values in the NumericVector x sorted from smallest to largest.
}
\description{
Sort the values in a NumericVector using Rcpp
}
\examples{
\dontrun{
A <- c(2,1,4,5)
SortC(A)
}
}
