\name{StatComp22058-package}
\alias{StatComp22058-package}
\alias{StatComp22058}
\docType{package}
\title{
  KCP(Kernel Change Point) permutation test
}
\description{
  Several R and Rcpp function are included in this package to implement the KCP permutation test.
}
\details{
  There are two main functions, namely VarKCP for variance test and Vardrop for variance drop test.
}
\author{
22058
Maintainer: 22058 <gg05070514@163.com>
}
\references{
  Cabrieto J, Tuerlinckx F, Kuppens P, et al. Testing for the Presence of Correlation Changes in a Multivariate Time Series: A Permutation Based Approach[J]. Scientific Reports, 2018, 8(1).
}
\keyword{ package }
