% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VexpressionC}
\alias{VexpressionC}
\title{The within-phase scatter using Rcpp}
\usage{
VexpressionC(X, n, a, b)
}
\arguments{
\item{X}{a matrix contains multivariate time series data. The rows are time points and the columns are variables.}

\item{n}{the number of time points}

\item{a}{the last observation in the last phase}

\item{b}{the last observation in the current phase}
}
\value{
the value of within-phase scatter for the current phase
}
\description{
Compute the within-phase scatter using Rcpp
}
\examples{
\dontrun{
data(mydata)
attach(mydata)
N <- length(mydata[,1])
VC <- VexpressionC(cordata,N,0,N)
VC
}
}
