% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCPpermR.R
\name{VexpressionR}
\alias{VexpressionR}
\title{The within-phase scatter using R}
\usage{
VexpressionR(X, n, a, b)
}
\arguments{
\item{X}{a matrix contains multivariate time series data. The rows are time points and the columns are variables.}

\item{n}{the number of time points}

\item{a}{the last observation in the last phase}

\item{b}{the last observation in the current phase}
}
\value{
the value of within-phase scatter for the current phase
}
\description{
Compute the within-phase scatter using R
}
\examples{
\dontrun{
data(mydata)
attach(mydata)
N <- length(mydata[,1])
VR <- VexpressionR(cordata,N,0,N)
VR
}
}
